Primitive entity player-selection in ScEd
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@1727 3db68df2-c116-0410-a063-a993310a9797

+++ source/tools/sced/CommandManager.cpp
@@ -38,7 +38,7 @@ void CCommandManager::Append(CCommand* cmd)
 		}
 
 		// update history position to point at last command executed
-		m_HistoryPos=m_CommandHistory.size()-1;
+		m_HistoryPos=(i32)m_CommandHistory.size()-1;
 	}
 }
